rework library loading
Previously the library was loaded with a path specified at compiletime,
which made it impossible to put the executable to a different system
without having the library in the exact same path. This can either be
cumbersome or impossible in case of a linux - windows crosscompile.

This reworked loading now uses the `library_path` from the builder. The
builder still uses the compiletime path as default, so nothing should
break.

This also removes another dependency `lazy_static` as a sideresult.